Change 'const long' -> 'long' in return types.
The const is ignored here.

+++ libhmsbeagle/beagle.h
@@ -139,53 +139,53 @@ enum BeagleReturnCodes {
  * This enumerates all possible hardware and implementation capability flags.
  * Each capability is a bit in a 'long'
  */
-enum BeagleFlags : long {
-    BEAGLE_FLAG_PRECISION_SINGLE    = 1UL << 0,    /**< Single precision computation */
-    BEAGLE_FLAG_PRECISION_DOUBLE    = 1UL << 1,    /**< Double precision computation */
+enum BeagleFlags : long long {
+    BEAGLE_FLAG_PRECISION_SINGLE    = 1ULL << 0,    /**< Single precision computation */
+    BEAGLE_FLAG_PRECISION_DOUBLE    = 1ULL << 1,    /**< Double precision computation */
 
-    BEAGLE_FLAG_COMPUTATION_SYNCH   = 1UL << 2,    /**< Synchronous computation (blocking) */
-    BEAGLE_FLAG_COMPUTATION_ASYNCH  = 1UL << 3,    /**< Asynchronous computation (non-blocking) */
+    BEAGLE_FLAG_COMPUTATION_SYNCH   = 1ULL << 2,    /**< Synchronous computation (blocking) */
+    BEAGLE_FLAG_COMPUTATION_ASYNCH  = 1ULL << 3,    /**< Asynchronous computation (non-blocking) */
 
-    BEAGLE_FLAG_EIGEN_REAL          = 1UL << 4,    /**< Real eigenvalue computation */
-    BEAGLE_FLAG_EIGEN_COMPLEX       = 1UL << 5,    /**< Complex eigenvalue computation */
+    BEAGLE_FLAG_EIGEN_REAL          = 1ULL << 4,    /**< Real eigenvalue computation */
+    BEAGLE_FLAG_EIGEN_COMPLEX       = 1ULL << 5,    /**< Complex eigenvalue computation */
 
-    BEAGLE_FLAG_SCALING_MANUAL      = 1UL << 6,    /**< Manual scaling */
-    BEAGLE_FLAG_SCALING_AUTO        = 1UL << 7,    /**< Auto-scaling on (deprecated, may not work correctly) */
-    BEAGLE_FLAG_SCALING_ALWAYS      = 1UL << 8,    /**< Scale at every updatePartials (deprecated, may not work correctly) */
-    BEAGLE_FLAG_SCALING_DYNAMIC     = 1UL << 25,   /**< Manual scaling with dynamic checking (deprecated, may not work correctly) */
+    BEAGLE_FLAG_SCALING_MANUAL      = 1ULL << 6,    /**< Manual scaling */
+    BEAGLE_FLAG_SCALING_AUTO        = 1ULL << 7,    /**< Auto-scaling on (deprecated, may not work correctly) */
+    BEAGLE_FLAG_SCALING_ALWAYS      = 1ULL << 8,    /**< Scale at every updatePartials (deprecated, may not work correctly) */
+    BEAGLE_FLAG_SCALING_DYNAMIC     = 1ULL << 25,   /**< Manual scaling with dynamic checking (deprecated, may not work correctly) */
 
-    BEAGLE_FLAG_SCALERS_RAW         = 1UL << 9,    /**< Save raw scalers */
-    BEAGLE_FLAG_SCALERS_LOG         = 1UL << 10,   /**< Save log scalers */
+    BEAGLE_FLAG_SCALERS_RAW         = 1ULL << 9,    /**< Save raw scalers */
+    BEAGLE_FLAG_SCALERS_LOG         = 1ULL << 10,   /**< Save log scalers */
 
-    BEAGLE_FLAG_INVEVEC_STANDARD    = 1UL << 20,   /**< Inverse eigen vectors passed to BEAGLE have not been transposed */
-    BEAGLE_FLAG_INVEVEC_TRANSPOSED  = 1UL << 21,   /**< Inverse eigen vectors passed to BEAGLE have been transposed */
+    BEAGLE_FLAG_INVEVEC_STANDARD    = 1ULL << 20,   /**< Inverse eigen vectors passed to BEAGLE have not been transposed */
+    BEAGLE_FLAG_INVEVEC_TRANSPOSED  = 1ULL << 21,   /**< Inverse eigen vectors passed to BEAGLE have been transposed */
 
-    BEAGLE_FLAG_VECTOR_SSE          = 1UL << 11,   /**< SSE computation */
-    BEAGLE_FLAG_VECTOR_AVX          = 1UL << 24,   /**< AVX computation */
-    BEAGLE_FLAG_VECTOR_NONE         = 1UL << 12,   /**< No vector computation */
+    BEAGLE_FLAG_VECTOR_SSE          = 1ULL << 11,   /**< SSE computation */
+    BEAGLE_FLAG_VECTOR_AVX          = 1ULL << 24,   /**< AVX computation */
+    BEAGLE_FLAG_VECTOR_NONE         = 1ULL << 12,   /**< No vector computation */
 
-    BEAGLE_FLAG_THREADING_CPP       = 1UL << 30,   /**< C++11 threading */
-    BEAGLE_FLAG_THREADING_OPENMP    = 1UL << 13,   /**< OpenMP threading */
-    BEAGLE_FLAG_THREADING_NONE      = 1UL << 14,   /**< No threading (default) */
+    BEAGLE_FLAG_THREADING_CPP       = 1ULL << 30,   /**< C++11 threading */
+    BEAGLE_FLAG_THREADING_OPENMP    = 1ULL << 13,   /**< OpenMP threading */
+    BEAGLE_FLAG_THREADING_NONE      = 1ULL << 14,   /**< No threading (default) */
 
-    BEAGLE_FLAG_PROCESSOR_CPU       = 1UL << 15,   /**< Use CPU as main processor */
-    BEAGLE_FLAG_PROCESSOR_GPU       = 1UL << 16,   /**< Use GPU as main processor */
-    BEAGLE_FLAG_PROCESSOR_FPGA      = 1UL << 17,   /**< Use FPGA as main processor */
-    BEAGLE_FLAG_PROCESSOR_CELL      = 1UL << 18,   /**< Use Cell as main processor */
-    BEAGLE_FLAG_PROCESSOR_PHI       = 1UL << 19,   /**< Use Intel Phi as main processor */
-    BEAGLE_FLAG_PROCESSOR_OTHER     = 1UL << 26,   /**< Use other type of processor */
+    BEAGLE_FLAG_PROCESSOR_CPU       = 1ULL << 15,   /**< Use CPU as main processor */
+    BEAGLE_FLAG_PROCESSOR_GPU       = 1ULL << 16,   /**< Use GPU as main processor */
+    BEAGLE_FLAG_PROCESSOR_FPGA      = 1ULL << 17,   /**< Use FPGA as main processor */
+    BEAGLE_FLAG_PROCESSOR_CELL      = 1ULL << 18,   /**< Use Cell as main processor */
+    BEAGLE_FLAG_PROCESSOR_PHI       = 1ULL << 19,   /**< Use Intel Phi as main processor */
+    BEAGLE_FLAG_PROCESSOR_OTHER     = 1ULL << 26,   /**< Use other type of processor */
 
-    BEAGLE_FLAG_FRAMEWORK_CUDA      = 1UL << 22,   /**< Use CUDA implementation with GPU resources */
-    BEAGLE_FLAG_FRAMEWORK_OPENCL    = 1UL << 23,   /**< Use OpenCL implementation with GPU resources */
-    BEAGLE_FLAG_FRAMEWORK_CPU       = 1UL << 27,   /**< Use CPU implementation */
+    BEAGLE_FLAG_FRAMEWORK_CUDA      = 1ULL << 22,   /**< Use CUDA implementation with GPU resources */
+    BEAGLE_FLAG_FRAMEWORK_OPENCL    = 1ULL << 23,   /**< Use OpenCL implementation with GPU resources */
+    BEAGLE_FLAG_FRAMEWORK_CPU       = 1ULL << 27,   /**< Use CPU implementation */
 
-    BEAGLE_FLAG_PARALLELOPS_STREAMS = 1UL << 28,   /**< Operations in updatePartials may be assigned to separate device streams */
-    BEAGLE_FLAG_PARALLELOPS_GRID    = 1UL << 29,   /**< Operations in updatePartials may be folded into single kernel launch (necessary for partitions; typically performs better for problems with fewer pattern sites) */
+    BEAGLE_FLAG_PARALLELOPS_STREAMS = 1ULL << 28,   /**< Operations in updatePartials may be assigned to separate device streams */
+    BEAGLE_FLAG_PARALLELOPS_GRID    = 1ULL << 29,   /**< Operations in updatePartials may be folded into single kernel launch (necessary for partitions; typically performs better for problems with fewer pattern sites) */
 
-    BEAGLE_FLAG_PREORDER_TRANSPOSE_MANUAL = 1UL << 30, /**< Pre-order transition matrices passed to BEAGLE have been transposed */
-    BEAGLE_FLAG_PREORDER_TRANSPOSE_AUTO   = 1UL << 31, /**< Automatically transpose pre-order transition matrices */
+    BEAGLE_FLAG_PREORDER_TRANSPOSE_MANUAL = 1ULL << 30, /**< Pre-order transition matrices passed to BEAGLE have been transposed */
+    BEAGLE_FLAG_PREORDER_TRANSPOSE_AUTO   = 1ULL << 31, /**< Automatically transpose pre-order transition matrices */
 
-    BEAGLE_FLAG_COMPUTATION_ACTION        = 1UL << 40    /**< Action computation */
+    BEAGLE_FLAG_COMPUTATION_ACTION        = 1ULL << 40    /**< Action computation */
 };
 
 
@@ -226,7 +226,7 @@ typedef struct {
     char* implName;     /**< Name of implementation on which this instance is running as a
                          *   NULL-terminated character string */
     char* implDescription; /**< Description of implementation with details such as how auto-scaling is performed */
-    long flags;         /**< Bit-flags that characterize the activate
+    long long flags;    /**< Bit-flags that characterize the activate
                          *   capabilities of the resource and implementation for this instance */
 } BeagleInstanceDetails;
 
@@ -236,7 +236,7 @@ typedef struct {
 typedef struct {
     char* name;         /**< Name of resource as a NULL-terminated character string */
     char* description;  /**< Description of resource as a NULL-terminated character string */
-    long  supportFlags; /**< Bit-flags of supported capabilities on resource */
+    long long  supportFlags; /**< Bit-flags of supported capabilities on resource */
     long  requiredFlags;/**< Bit-flags that identify resource type */
 } BeagleResource;
 
@@ -255,7 +255,7 @@ typedef struct {
     int number;         /**< Resource number  */
     char* name;         /**< Name of resource as a NULL-terminated character string */
     char* description;  /**< Description of resource as a NULL-terminated character string */
-    long  supportFlags; /**< Bit-flags of supported capabilities on resource */
+    long long  supportFlags; /**< Bit-flags of supported capabilities on resource */
     long  requiredFlags;/**< Bit-flags that identify resource type */
     int   returnCode;   /**< Return code of for benchmark attempt (see BeagleReturnCodes) */
     char* implName;     /**< Name of implementation used to benchmark resource */
